rename new map component
