+++ src/StamenMap/Shapes.js
