+++ src/StamenMap/StamenMap.test.js
